fix: update executor configuration to use environment variables for connection and table settings; refactor Executor class for improved structure and error handling
