+++ config/executor.php
@@ -4,10 +4,10 @@ return [
     /**
      * The database connection that should be used by the executor.
      */
-    'connection' => env('DB_CONNECTION', config('database.default')),
+    'connection' => env('EXECUTOR_CONNECTION', config('database.default')),
 
     /**
      * The table that should be used to store the executors.
      */
-    'table'      => 'executors',
+    'table' => env('EXECUTOR_TABLE', 'executors'),
 ];
